Revert "feat: add sorting by column type on repositories screen (#662)"
This reverts commit d5880321ed82d5c505c74c6f797e70d5895803b6.
