+++ components/atoms/TableTitle/table-title.tsx
@@ -3,13 +3,11 @@ import React from "react";
 interface TableTitleProps {
     text: string;
     className?: string;
-    handleClick?: () => void;
 }
 
-const TableTitle: React.FC<TableTitleProps> = ({ className, text, handleClick }) => {
-
+const TableTitle: React.FC<TableTitleProps> = ({ className, text }) => {
   return (
-    <div className={`${className ? className : ""} font-semibold tracking-wide text-xs text-light-slate-11 uppercase`} onClick={handleClick}>
+    <div className={`${className ? className : ""} font-semibold tracking-wide text-xs text-light-slate-11 uppercase`}>
       {text}
     </div>
   );
