+++ lib/hooks/useRepositoriesList.ts
@@ -16,6 +16,7 @@ const useRepositoriesList = (skipFilters = false, repoIds: number[] = []) => {
   const [page, setPage] = useState(1);
   const [limit, setLimit] = useState(10);
   const range = useStore(state => state.range);
+
   const baseEndpoint = !skipFilters ? `${topic}/repos` : "repos/list";
   const pageQuery = page ? `page=${page}` : "";
   const filterQuery = getFilterQuery(selectedFilter);
